refactor: separate the  type to make it reusable
